1.优化支付功能 2.手动实现三级联动 3.支付成功后的插入数据

+++ src/main/java/com/eight/controller/manage/user/MUserManageController.java
@@ -83,7 +83,7 @@ public class MUserManageController {
      * 通过id查询信息
      */
     @RequestMapping("/queryUserById")
-    public String queryUserById(int userId,Model model){
+    public String queryUserById(Long userId,Model model){
         User user = mUserService.queryUserById(userId);
         model.addAttribute("obj", user);
         return "manage/user/userPage";
